Make vendored WifiWhirl library buildable in ESPHome (no subdir resources)
