+++ esphome/components/wifiwhirl_bwc_unified/__init__.py
@@ -1,0 +1,2 @@
+# Helper component: ships the vendored WifiWhirl BWC_unified library sources
+CODEOWNERS = ["@Finest"]
