+++ esphome/components/wifiwhirl_bwc_unified/enums.h
