+++ esphome/components/wifiwhirl_bwc_unified/pitches.h
