+++ esphome/components/wifiwhirl_cio/CIO_2021HJT.h
@@ -1,9 +1,9 @@
 #pragma once
 #include <Arduino.h>
-#include "../BWC_unified/enums.h"
+#include "esphome/components/wifiwhirl_bwc_unified/enums.h"
 #include "CIO_TYPE1.h"
 
-class CIO_2021 : public CIO_6_TYPE1
+class CIO_2021_HJT : public CIO_6_TYPE1
 {
     /*These must be declared for the API to work*/
 public:
@@ -19,11 +19,11 @@ public:
 
     /*These must be declared for the API to work*/
 public:
-    const String _MYMODEL = "MIAMI2021";
-    const bool _HASJETS = false;
+    const String _MYMODEL = "MALDIVES2021";
+    const bool _HASJETS = true;
     const bool _HASAIR = true;
 
 private:
     /*Binary representation of button presses on the display, like 0x1B1B.*/
-    uint16_t button_codes[10] = {0x1B1B, 0x0100, 0x0300, 0x1212, 0x0809, 0x1012, 0x1112, 0x1312, 0x0200, 0x0000};
+    uint16_t button_codes[11] = {0x1B1B, 0x0100, 0x0300, 0x1212, 0x0a09, 0x1012, 0x1312, 0x0809, 0x0200, 0x0000, 0x1112};
 };
