+++ esphome/components/wifiwhirl_cio/CIO_BASE.h
@@ -1,6 +1,6 @@
 #pragma once
 #include <Arduino.h>
-#include "../BWC_unified/enums.h"
+#include "esphome/components/wifiwhirl_bwc_unified/enums.h"
 
 constexpr int MAXBUTTONS = 50;
