+++ esphome/components/wifiwhirl_cio/CIO_PRE2021.cpp
