+++ esphome/components/wifiwhirl_cio/CIO_PRE2021.h
@@ -1,6 +1,6 @@
 #pragma once
 #include <Arduino.h>
-#include "../BWC_unified/enums.h"
+#include "esphome/components/wifiwhirl_bwc_unified/enums.h"
 #include "CIO_TYPE1.h"
 
 class CIO_PRE2021 : public CIO_6_TYPE1
