+++ esphome/components/wifiwhirl_cio/__init__.py
@@ -1,0 +1,2 @@
+# Helper component: ships the vendored WifiWhirl CIO library sources
+CODEOWNERS = ["@Finest"]
