+++ esphome/components/wifiwhirl_dsp/__init__.py
@@ -1,0 +1,2 @@
+# Helper component: ships the vendored WifiWhirl DSP library sources
+CODEOWNERS = ["@Finest"]
